requirements.txt
pytesseract~=0.3.10
numpy~=1.26.4
opencv-python~=4.9.0.80
python-dotenv~=1.0.1
selenium~=4.21.0
pillow~=10.3.0
pycryptodome~=3.20.0
psycopg2~=2.9.9